+++ config/initializers/exception_notification.rb
@@ -14,11 +14,19 @@ IGNORED_EXCEPTIONS = [
   'ActionController::UnknownHttpMethod'
 ].freeze
 
+# Check if the user configured a special address to send
+# the notifications from. If not, use notifier@domain
+if Figaro.env.exception_notifier_sender.present?
+  sender = Figaro.env.exception_notifier_sender
+else
+  sender = "notifier@#{Figaro.env.application_domain!}"
+end
+
 # Otherwise, setup the exception notifier
 Rails.application.config.middleware.use ExceptionNotification::Rack,
                                         ignore_exceptions: IGNORED_EXCEPTIONS + ExceptionNotifier.ignored_exceptions,
                                         email:             {
                                           email_prefix:         '[EXCEPTION NOTIFICATION] ',
-                                          sender_address:       "notifier@#{Figaro.env.application_domain!}",
+                                          sender_address:       sender,
                                           exception_recipients: Figaro.env.exception_notifier_recipient
                                         }